Clear up the enums and fix implicit condition-bug
The OUT_F-type was not clear. Make it easier to see the out-file
has no state (=NONE) and is defined for each slot individually.

Moreover, in the initial creation, the err-file fell through
just because STATIC = 0 and .outfile wasn't defined for it.

It was only coincidence this worked and now we have a much
more bulletproof implementation.