Don't lock up on blocking call_out
Since the av-refactor, we keep a call active as long as there is
a listener or a sender.
In case we don't have a listener, don't lock up in the sending
function and rather just fall-through in case the write blocks.
We can't do that as easily with the file-data-code, because data-
integrity is very important there.