Remove useless call to FD_CLR()
This snippet was introduced when the client could send too much data to
the core for the transfer.
The new toxcore API introduces a callback triggered when the core is
ready to send more data. It means we don't need to keep track of a
local buffer anymore, and simply read from the file_in FIFO when the
core is ready.