Get audio call to work properly
This commit puts back in place the settings that were previously set
regarding audio encoding, preparation of the frame, buffer and so on.
These settings used to work great and I modified them without having any
clue of how it was working.

Now I'm done playing, and it works properly. Hooray.