Allocate resources for outgoing call when needed
If we're not initiating the call, we might be in TRANSMITTING mode even
if call_in hasn't been openned by the user. In this case, we shouldn't
send any data.
sednfriendcalldata() should only be called when resources are allocated
for an OUTGOING call, which is triggered when the friend accepts audio.