Fix warnings.
The title of a conference is apparently not known immediately after joining the
conference, so don't print a warning in confcreate(), just wait for the
cbconftitle() callback immediately afterwards.
Friends that are not online can't be invited to conferences, logmsg' in that case.
Print the name of the conference before messages in cbconfmessage().