Major error message overhaul.
eprintf() is used with only the function name that failed
(and possibly a parameter) and then exits. Example:
eprintf("read %s:", f->name);

weprintf() is used when another function failed, but the
program does not exit. The error message is a sentence.
Example: weprintf("No friend with id %s found for %s\n", buf, c->numstr);

logmsg() is not used when anything failed, it is just for
displaying useful information for the user. There should be
three logmsg formats:

The first format consists of the area the logmsg() is concerned
with, for example "Name", followed by a ">" sign, and then
followed by the change within that area.
This format deals with non-group and non-friend specific changes.
Example: logmsg("State > %s\n", buf);

The second format displays friend-specific information. It
consists of a colon, followed by the friend name, followed by
another colon, followed by the value that was changed and then
the change itself.
Example: logmsg(": %s : Rx > Accepted\n", f->name);

The third format is similar to the second one, it consists of
a dash ("-"), followed by a conference number, a colon,
followed by the value that was changed and then the change itself.
This format tells the user about conference-specific changes.
Example: logmsg("- %s : Title > %s\n", c-numstr, title);

Errors when trying to change something in the gslots are written
with dprintf into the gslots .fd[ERR] and with weprintf() on
the screen.

This commit and the last one are attempts at implementing these rules.